changed: energies are now double instead of float

+++ include/BALL/MOLMEC/CHARMM/charmmBend.h
@@ -1,4 +1,4 @@
-// $Id: charmmBend.h,v 1.2 2000/02/14 22:43:57 oliver Exp $
+// $Id: charmmBend.h,v 1.3 2000/03/26 12:47:34 oliver Exp $
 // Molecular Mechanics: Charmm force field, bond stretch component
 
 #ifndef BALL_MOLMEC_CHARMM_CHARMMBEND_H
@@ -67,7 +67,7 @@ namespace BALL
 
 		/**	Calculates and returns the component's energy.
 		*/
-		virtual float updateEnergy();
+		virtual double updateEnergy();
 
 		/**	Calculates and returns the component's forces.
 		*/
