+++ source/MOLMEC/AMBER/amberTorsion.C
@@ -1,4 +1,4 @@
-// $Id: amberTorsion.C,v 1.17 2000/02/14 22:44:06 oliver Exp $
+// $Id: amberTorsion.C,v 1.18 2000/03/26 12:52:25 oliver Exp $
 
 #include <BALL/MOLMEC/AMBER/amberTorsion.h>
 #include <BALL/MOLMEC/AMBER/amber.h>
@@ -350,9 +350,9 @@ namespace BALL
 	}
 
 	// calculates the current energy of this component
-	float AmberTorsion::updateEnergy() 
+	double AmberTorsion::updateEnergy() 
 	{
-		float cosphi;
+		double cosphi;
 
 		Vector3	a21;
 		Vector3 a23;
@@ -378,8 +378,8 @@ namespace BALL
 				cross2321 = a23 % a21;
 				cross2334 = a23 % a34;
 
-				float length_cross2321 = cross2321.getLength();
-				float length_cross2334 = cross2334.getLength();
+				double length_cross2321 = cross2321.getLength();
+				double length_cross2334 = cross2334.getLength();
 
 				if (length_cross2321 != 0 && length_cross2334 != 0) 
 				{
@@ -408,8 +408,8 @@ namespace BALL
 	// calculates and adds its forces to the current forces of the force field
 	void AmberTorsion::updateForces()
 	{
-		float cosphi;
-		float dEdphi;
+		double cosphi;
+		double dEdphi;
 
 		Vector3	ab;	// vector from atom2 to atom1
 		Vector3 cb;		// vector from atom2 to atom3
@@ -425,23 +425,23 @@ namespace BALL
 					 || it->atom3->isSelected() || it->atom4->isSelected())))
 			{
 				ab = it->atom1->getPosition() - it->atom2->getPosition();
-				float length_ab = ab.getLength();
+				double length_ab = ab.getLength();
 				Vector3 ba = it->atom2->getPosition() - it->atom1->getPosition();
 				cb = it->atom3->getPosition() - it->atom2->getPosition();
-				float length_cb = cb.getLength();
+				double length_cb = cb.getLength();
 				dc = it->atom4->getPosition() - it->atom3->getPosition();
-				float length_dc = dc.getLength();
+				double length_dc = dc.getLength();
 
 				if (length_ab != 0 && length_cb != 0 && length_dc != 0) 
 				{
 					Vector3  t = ba % cb;   // cross product of cb and ba
 					Vector3  u = cb % dc;   // cross product of cb and dc
 
-					float length_t2 = t.getSquareLength();
-					float length_u2 = u.getSquareLength();
+					double length_t2 = t.getSquareLength();
+					double length_u2 = u.getSquareLength();
 
-					float length_t = sqrt(length_t2);
-					float length_u = sqrt(length_u2);
+					double length_t = sqrt(length_t2);
+					double length_u = sqrt(length_u2);
 
 					if (length_t != 0 && length_u != 0) 
 					{
@@ -462,7 +462,7 @@ namespace BALL
 						dEdphi = (-it->V) * (1e13 / Constants::AVOGADRO) * it->f * sin(it->f * acos(cosphi) - it->phase);
 
 
-						float direction = (t % u) * cb;
+						double direction = (t % u) * cb;
 						if (direction > 0.0)
 						{
 							dEdphi = -dEdphi;
@@ -470,8 +470,8 @@ namespace BALL
 
 						Vector3 ca = it->atom3->getPosition() - it->atom1->getPosition();
 						Vector3 db = it->atom4->getPosition() - it->atom2->getPosition();
-						Vector3 dEdt =   dEdphi / (length_t2 * cb.getLength()) * (t % cb);
-						Vector3 dEdu = - dEdphi / (length_u2 * cb.getLength()) * (u % cb);
+						Vector3 dEdt =   (float)(dEdphi / (length_t2 * cb.getLength())) * (t % cb);
+						Vector3 dEdu = - (float)(dEdphi / (length_u2 * cb.getLength())) * (u % cb);
 	
 
 						if (getForceField()->getUseSelection() == false)
@@ -480,7 +480,9 @@ namespace BALL
 							it->atom2->getForce() += ca % dEdt + dEdu % dc;
 							it->atom3->getForce() += dEdt % ba + db % dEdu;
 							it->atom4->getForce() += dEdu % cb; 
-						} else {
+						} 
+						else 
+						{
 							if (it->atom1->isSelected()) it->atom1->getForce() += dEdt % cb;
 							if (it->atom2->isSelected()) it->atom2->getForce() += ca % dEdt + dEdu % dc;
 							if (it->atom3->isSelected()) it->atom3->getForce() += dEdt % ba + db % dEdu;
